18 pass

+++ homework_02/engine.py
@@ -3,5 +3,5 @@ from dataclasses import dataclass
 
 @dataclass()
 class Engine():
-    volume: float
+    volume: int
     pistons: int
